changes in responsibilities  in the components
